# Remove some allocations related to storing CacheEntry scopes