While looking at the source code to address #45436 I've realized that the `_previous` field of the `CacheEntryStack` type is never used:

https://github.com/dotnet/runtime/blob/44870889827a8551e6d71f089842085a64c88527/src/libraries/Microsoft.Extensions.Caching.Memory/src/CacheEntryStack.cs#L8-L38

This led me to the conclusion, that there is no need to have this type if all it does is just being a wrapper around a single `CacheEntry` instance. So instead of `AsyncLocal<CacheEntryStack>` we can have `AsyncLocal<CacheEntry>`

The `ScopeLease` responsibility was to just reset the `Current` during Dispose. To save some allocation we can just "inline" this type and do what it did in `CacheEntry.Dispose`:

https://github.com/dotnet/runtime/blob/44870889827a8551e6d71f089842085a64c88527/src/libraries/Microsoft.Extensions.Caching.Memory/src/CacheEntryHelper.cs#L59-L62

Doing this a little bit later than before allowed to avoid expensive access to `CacheEntryHelper.Current` in `CacheEntry.Dispose`

Benchmark numbers:

|                           Method |           Toolchain |     Mean | Ratio | Allocated |
|--------------------------------- |-------------------- |---------:|------:|----------:|
|       AddThenRemove_NoExpiration |  \after\CoreRun.exe | 59.09 us |  0.94 |     57 KB |
|       AddThenRemove_NoExpiration | \before\CoreRun.exe | 63.14 us |  1.00 |     71 KB |
|                                  |                     |          |       |           |
| AddThenRemove_AbsoluteExpiration |  \after\CoreRun.exe | 54.01 us |  0.80 |     45 KB |
| AddThenRemove_AbsoluteExpiration | \before\CoreRun.exe | 67.33 us |  1.00 |     71 KB |
|                                  |                     |          |       |           |
| AddThenRemove_RelativeExpiration |  \after\CoreRun.exe | 60.80 us |  0.89 |     57 KB |
| AddThenRemove_RelativeExpiration | \before\CoreRun.exe | 68.03 us |  1.00 |     71 KB |
|                                  |                     |          |       |           |
|  AddThenRemove_SlidingExpiration |  \after\CoreRun.exe | 58.53 us |  0.88 |     60 KB |
|  AddThenRemove_SlidingExpiration | \before\CoreRun.exe | 66.32 us |  1.00 |     72 KB

Contributes to #45436