@adamsitnik 2020-12-11
> The ordering change also concerns me but I haven't spend enough time coming up with hypothetical scenarios it might affect.

@davidfowl thank you for your feedback, I've sent https://github.com/dotnet/runtime/pull/45964 to address it.